requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.12.0
    # via -r requirements.in
numpy==1.20.3
    # via opencv-python
opencv-python==4.5.2.54
    # via -r requirements.in
six==1.16.0
    # via absl-py
